Reuse existing task definitions
Unless you provide your own task definition in dagster.yaml, every run
that the EcsRunLauncher launches registers a new task definition.

For most runs, this is redundant - the new task definition it creates is
identical to the last one it created. Consequently:

- Users' AWS accounts are slowly filling up with redundant ECS Task
  Definitions
- There's a hard limit of one million revisions per task definition family
  so the first user to hit a million runs on the EcsRunLauncher has a
  nasty ticking time bomb waiting for them

This change aims to alter that behavior by treating the task definition
as a more generic canvas and leaning into Task Overides and Continer
Overrides to vary actual behavior. As long as we have a minimum
sufficient task definition, we should be able to reuse it and customize
it via overrides:

https://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_TaskOverride.html
https://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_ContainerOverride.html